feat: support GPU-calculated luts in vispy (#77)
* feat: support gpu clims

* Update _utils.py


* fix downcast

* style(pre-commit.ci): auto fixes [...]

* fix for int16

* style(pre-commit.ci): auto fixes [...]

---------

+++ examples/zarr_arr.py
@@ -12,4 +12,4 @@ except ImportError:
 URL = "https://s3.embl.de/i2k-2020/ngff-example-data/v0.4/tczyx.ome.zarr"
 zarr_arr = zarr.open(URL, mode="r")
 
-ndv.imshow(zarr_arr["s0"])
+ndv.imshow(zarr_arr["s0"].astype("uint16"))
